requirements.txt
absl-py==1.4.0
aiodns==3.0.0
aiohttp==3.8.4
aiortsp==1.3.6
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
av==10.0.0
Brotli==1.0.9
cachetools==5.3.0
cchardet==2.1.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.0.4
contourpy==1.0.7
customtkinter==5.1.3
cycler==0.11.0
Cython==0.29.34
darkdetect==0.8.0
decorator==5.1.1
deep-sort-realtime==1.3.2
docopt==0.6.2
dpkt==1.9.8
emoji==2.2.0
filelock==3.12.0
fire==0.5.0
flatbuffers==23.5.8
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.3.0
google-api-core==2.11.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-cloud-vision==3.4.1
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.48.2
huggingface-hub==0.14.1
idna==3.4
ifaddr==0.2.0
imageio==2.30.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
kiwisolver==1.4.4
numpy==1.24.3
lazy_loader==0.2
llvmlite==0.40.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mediapipe==0.9.1.0
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
networkx==3.1
numba==0.57.0
oauthlib==3.2.2
openai==0.27.6
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.1
pafy==0.5.5
pandas==1.5.3
Pillow==9.5.0
pipreqs==0.4.13
playsound==1.3.0
proto-plus==1.22.2
protobuf==3.20.3
psutil==5.9.5
pupil-labs-realtime-api==1.1.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybboxes==0.1.6
pycares==4.3.0
pycparser==2.21
pynput==1.7.6
pyobjc==9.0.1
pyparsing==3.0.9
pyperclip==1.8.2
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
ratelim==0.1.6
regex==2023.5.5
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
sahi==0.11.13
scikit-image==0.20.0
scipy==1.9.1
screeninfo==0.8.1
seaborn==0.12.2
sentry-sdk==1.22.2
shapely==2.0.1
six==1.16.0
sounddevice==0.4.6
tensorboard==2.13.0
tensorboard-data-server==0.7.0
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
tifffile==2023.4.12
tiktoken==0.3.3
tokenizers==0.13.3
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
transformers==4.26.1
transitions==0.9.0
ttkbootstrap==1.10.1
typing_extensions==4.5.0
ultralytics==8.0.43
ultralyticsplus==0.0.28
urllib3==1.26.15
wavio==0.0.7
websockets==11.0.3
Werkzeug==2.3.4
yarg==0.1.9
yarl==1.9.2
youtube-dl==2020.12.2
zeroconf==0.62.0
zipp==3.15.0
torchaudio==0.13.1
